refactor(telemetry): extract framerate tracking to separate file
Moves framerate measurement logic into framerateTracker.ts and adds a
TRACK_REACT_FRAMERATE feature flag as an alternative gate alongside the
existing 10% session sampling.
