+++ frontend/src/framerateTracker.ts
@@ -1,0 +1,102 @@
+import type { PostHog } from 'posthog-js'
+
+const LONG_FRAME_THRESHOLD_MS = 50
+const CAPTURE_INTERVAL_MS = 30_000
+
+export function startFramerateTracking(posthog: PostHog): void {
+    let rafId: number | null = null
+    let captureIntervalId: number | null = null
+    let previousTimestamp: number | null = null
+    let frameCount = 0
+    let frameTimeSum = 0
+    let shortestFrame = Infinity
+    let longestFrame = 0
+    let longFrameCount = 0
+    let measurementStart = 0
+
+    function reset(): void {
+        previousTimestamp = null
+        frameCount = 0
+        frameTimeSum = 0
+        shortestFrame = Infinity
+        longestFrame = 0
+        longFrameCount = 0
+        measurementStart = performance.now()
+    }
+
+    function capture(): void {
+        if (frameCount === 0) {
+            return
+        }
+        const elapsed = performance.now() - measurementStart
+        posthog.capture('react_framerate', {
+            avg_fps: Math.round((frameCount / elapsed) * 1000),
+            avg_frame_time_ms: Math.round((frameTimeSum / frameCount) * 100) / 100,
+            min_frame_time_ms: Math.round(shortestFrame * 100) / 100,
+            max_frame_time_ms: Math.round(longestFrame * 100) / 100,
+            long_frame_count: longFrameCount,
+            total_frames: frameCount,
+            measurement_duration_ms: Math.round(elapsed),
+        })
+        reset()
+    }
+
+    function onFrame(timestamp: number): void {
+        if (previousTimestamp !== null) {
+            const delta = timestamp - previousTimestamp
+            frameCount++
+            frameTimeSum += delta
+            if (delta < shortestFrame) {
+                shortestFrame = delta
+            }
+            if (delta > longestFrame) {
+                longestFrame = delta
+            }
+            if (delta > LONG_FRAME_THRESHOLD_MS) {
+                longFrameCount++
+            }
+        }
+        previousTimestamp = timestamp
+        rafId = requestAnimationFrame(onFrame)
+    }
+
+    function start(): void {
+        if (rafId !== null) {
+            return
+        }
+        reset()
+        rafId = requestAnimationFrame(onFrame)
+        captureIntervalId = window.setInterval(capture, CAPTURE_INTERVAL_MS)
+    }
+
+    function stop(): void {
+        if (rafId !== null) {
+            cancelAnimationFrame(rafId)
+            rafId = null
+        }
+        if (captureIntervalId !== null) {
+            clearInterval(captureIntervalId)
+            captureIntervalId = null
+        }
+    }
+
+    function onVisibilityChange(): void {
+        if (document.hidden) {
+            capture()
+            stop()
+        } else {
+            start()
+        }
+    }
+
+    document.addEventListener('visibilitychange', onVisibilityChange)
+
+    if (!document.hidden) {
+        start()
+    }
+
+    window.addEventListener('beforeunload', () => {
+        stop()
+        document.removeEventListener('visibilitychange', onVisibilityChange)
+    })
+}
